Creating a Controller for pratice
Controller can be thought as a manager in a restaurant.
Just as a manager directs different tasks to ensure the restaurant runs smoothly,
a controller in Laravel directs requests to the appropriate action to handle them.

Real like example: When we enter restaurant (make a web request ), the manager (controller) decides what will happen next. If we want a order food (access a page), the manager sends you to the waiter (a specifice method in the controller) to take the order.

php artisan make:controller NameofController
 this will creates a new controller class

Controllers have methods, which are like tasks the manager handles. Each method corresponds to a specific action.

Analogy: If a restaurant manager handles different tasks like welcoming guests, taking orders, and managing bills, a controller handles different requests like showing a homepage, displaying a product, or processing a form submission.